cloud_storage: break out configuration
Fixes gnarly issue in which the configuration was pulling in http client
dependencies all over.
