+++ src/v/archival/purger.h
@@ -16,6 +16,7 @@
 #include "cluster/types.h"
 
 #include <seastar/core/future.hh>
+#include <seastar/core/sharded.hh>
 
 namespace archival {
