+++ src/v/cluster/topic_table_partition_generator.h
@@ -12,6 +12,8 @@
 
 #include "cluster/topic_table.h"
 
+#include <seastar/core/sharded.hh>
+
 namespace cluster {
 struct partition_replicas {
     model::ntp partition;
