+++ src/v/cluster/topics_frontend.h
@@ -30,7 +30,10 @@
 
 #include <absl/container/flat_hash_map.h>
 
+#include <chrono>
 #include <system_error>
+
+using namespace std::chrono_literals;
 
 namespace cluster {
