+++ src/v/redpanda/monitor_unsafe_log_flag.cc
@@ -12,6 +12,7 @@
 #include "redpanda/monitor_unsafe_log_flag.h"
 
 #include "cluster/logger.h"
+#include "config/configuration.h"
 #include "strings/utf8.h"
 
 using cluster::clusterlog;
