+++ src/v/redpanda/tests/fixture.h
@@ -12,6 +12,7 @@
 #pragma once
 #include "archival/types.h"
 #include "cloud_roles/types.h"
+#include "cloud_storage/configuration.h"
 #include "cloud_storage_clients/configuration.h"
 #include "cluster/cluster_utils.h"
 #include "cluster/controller.h"
